setup the frontend and backend
