+++ client/src/App.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import './App.css';
+import {BrowserRouter, Link, Switch, Route} from 'react-router-dom'
+
+function App() {
+  return (
+    <BrowserRouter>
+      <div className="App">
+        <header className="App-header">
+          <p>
+            good luck guys.
+          </p>
+            Learn React
+        </header>
+      </div>
+    </BrowserRouter>
+  );
+}
+
+export default App;
